Added a thread tab 2 categories or this tab, one for data- and the other for resourcepack threads.
The Selection page uses a generic component that can be used in other places of the wiki too if needed.

The test threads are only placeholders and I didn't push any of that code for them yet.

(It is truely impressive how often I got disrupted while I was working on this. I could never work longer than about 30 minutes at a time.)